Split test_basic and test_ds_development
Create separate scripts for basic function testing and development. Test_basic now shows how to use the make_regular and BP_align methods
